backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv